+++ sdk/go/pulumi/context.go
@@ -235,6 +235,7 @@ func (ctx *Context) ReadResource(
 			Parent:     inputs.parent,
 			Properties: inputs.rpcProps,
 			Provider:   inputs.provider,
+			Id:         string(id),
 		})
 		if err != nil {
 			logging.V(9).Infof("RegisterResource(%s, %s): error: %v", t, name, err)