+++ CHANGELOG.md
@@ -5,8 +5,13 @@ CHANGELOG
 
 - Support passing a parent and providers for `ReadResource`, `RegisterResource`, and `Invoke` in the go SDK. [#3563](https://github.com/pulumi/pulumi/pull/3563)
 
+- Fix go SDK DeleteBeforeReplace [#3572](https://github.com/pulumi/pulumi/pull/3572)
+
 - Support for setting the `PULUMI_PREFER_YARN` environment variable to opt-in to using `yarn` instead of `npm` for
   installing Node.js dependencies. [#3556](https://github.com/pulumi/pulumi/pull/3556)
+
+- Fix regression that prevented relative paths passed to `--policy-pack` from working.
+  [#3565](https://github.com/pulumi/pulumi/issues/3564)
 
 ## 1.6.0 (2019-11-20)